# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Task Team Manager API is a backend service designed to manage team tasks using artificial intelligence.",
        "title": "AI Task Team Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.100.32:8080",
    "basePath": "/",
    "paths": {
...
        "/boards": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new board; the caller becomes the owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Create board",
                "parameters": [
                    {
                        "description": "Create board request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.CreateBoardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/board.BoardResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/boards/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get board by ID (user must be a member)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Get board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/board.BoardResponse"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "board not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/boards/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all members of a board with their employee profile info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Get board members",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/board.MemberResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/boards/{id}/statuses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns statuses of a board ordered by position; caller must be a member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Get board statuses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/board.StatusResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns user's boards and isFirstLogin flag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/board.DashboardResponse"
                        }
                    }
                }
            }
        },
...
        "/statuses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new status to a board; accepts title + board_id in body; caller must be a member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Create board status",
                "parameters": [
                    {
                        "description": "Create status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.CreateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/board.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/statuses/reorder": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update positions using board_status_id; caller must be a member of the board",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Reorder board statuses",
                "parameters": [
                    {
                        "description": "New order",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.ReorderStatusesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/statuses/{boardStatusId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a status from a board by board_status_id; caller must be a member",
                "tags": [
                    "Board"
                ],
                "summary": "Delete board status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board Status ID",
                        "name": "boardStatusId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "status not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update title and/or colour of a board status; caller must be a member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board"
                ],
                "summary": "Update board status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board Status ID",
                        "name": "boardStatusId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/board.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/board.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "status not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "board.BoardResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isOwner": {
                    "type": "boolean"
                },
                "memberCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "integer"
                }
            }
        },
        "board.CreateBoardRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "board.CreateStatusRequest": {
            "type": "object",
            "required": [
                "board_id",
                "title"
            ],
            "properties": {
                "board_id": {
                    "type": "integer"
                },
                "colour": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "board.DashboardResponse": {
            "type": "object",
            "properties": {
                "boards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/board.BoardResponse"
                    }
                },
                "isFirstLogin": {
                    "type": "boolean"
                }
            }
        },
        "board.MemberResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "board.ReorderStatusesRequest": {
            "type": "object",
            "required": [
                "statuses"
            ],
            "properties": {
                "statuses": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/board.StatusPosition"
                    }
                }
            }
        },
        "board.StatusPosition": {
            "type": "object",
            "required": [
                "board_status_id",
                "position"
            ],
            "properties": {
                "board_status_id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                }
            }
        },
        "board.StatusResponse": {
            "type": "object",
            "properties": {
                "board_status_id": {
                    "description": "id in board_statuses — use for reorder/delete",
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "colour": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
        "board.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "colour": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}